refactor: util 폴더에서 domain 과 연관된 함수 제거 - carRace 파일의 printCarRaceWinners 는 carRace 클래스 내부 메소드로 이동 - car 파일 이름을 cars 로 변경
